+++ cybsi/utils/converters.py
@@ -2,9 +2,16 @@
 A set of utility functions allowing to convert
 a string to a valid value of a chosen type.
 """
-from typing import Any, Callable, Dict
+from typing import Any, Callable, Dict, Type
 
-from cybsi.api.observable.enums import AttributeNames, EntityKeyTypes
+from cybsi.api.enum import CybsiAPIEnum
+from cybsi.api.observable.enums import (
+    AttributeNames,
+    EntityKeyTypes,
+    IdentityClass,
+    IndustrySector,
+    NodeRole,
+)
 
 
 def _str_converter(val: str):
@@ -29,6 +36,13 @@ def _bool_converter(val: str):
     if val.lower() == "false":
         return False
     raise ValueError("bool value must be 'true' or 'false'")
+
+
+def _new_enum_value_converter(api_enum: Type[CybsiAPIEnum], ignore_case=False):
+    def _enum_converter(val: Any) -> str:
+        return api_enum.from_string(str(val), ignore_case=ignore_case).value
+
+    return _enum_converter
 
 
 _entity_key_converters: Dict[EntityKeyTypes, Callable[[Any], Any]] = {
@@ -65,16 +79,16 @@ def convert_entity_key(k_type: EntityKeyTypes, val: Any) -> Any:
 
 _attr_value_converters: Dict[AttributeNames, Callable[[str], Any]] = {
     AttributeNames.Size: _int_converter,
-    AttributeNames.Class: _str_converter,
-    AttributeNames.Sectors: _str_converter,
     AttributeNames.DisplayNames: _str_converter,
     AttributeNames.Names: _str_converter,
-    AttributeNames.NodeRoles: _str_converter,
     AttributeNames.MalwareFamilyAliases: _str_converter,
     AttributeNames.IsIoC: _bool_converter,
     AttributeNames.IsTrusted: _bool_converter,
     AttributeNames.IsMalicious: _bool_converter,
     AttributeNames.IsDGA: _bool_converter,
+    AttributeNames.Class: _new_enum_value_converter(IdentityClass, ignore_case=True),
+    AttributeNames.NodeRoles: _new_enum_value_converter(NodeRole, ignore_case=True),
+    AttributeNames.Sectors: _new_enum_value_converter(IndustrySector, ignore_case=True),
 }